+++ src/repo/mongoose/model/clientEvent.ts
@@ -41,4 +41,9 @@ const schema = new mongoose.Schema(
     }
 );
 
-export default mongoose.model('ClientEvent', schema);
+export default mongoose.model('ClientEvent', schema)
+    .on('index', (error) => {
+        if (error !== undefined) {
+            console.error(error);
+        }
+    });