+++ src/repo/mongoose/model/gmoNotification.ts
@@ -47,4 +47,9 @@ const schema = new mongoose.Schema(
 // GMO売上健康診断時に使用
 schema.index({ jobCd: 1, tranDate: 1 });
 
-export default mongoose.model('GMONotification', schema);
+export default mongoose.model('GMONotification', schema)
+    .on('index', (error) => {
+        if (error !== undefined) {
+            console.error(error);
+        }
+    });